+++ src/pages/LoginPage/components/LoginForm.tsx
@@ -42,6 +42,8 @@ const LoginForm = () => {
     ) => {
       try {
         const response = await loginUser(loginData.email, loginData.password);
+        console.log(response,"login");
+        
         if (response && (response as { token?: string }).token) {
           setStatus({ success: true });
            navigate("/dashboard");   //SuperAdminDashboard
