+++ src/services/OrganizationService.ts
@@ -1,5 +1,5 @@
 import axios from "axios"
-import { API_URL } from "../constants/Constants"
+import { API_URL } from "../constants/Constants";
 
 export const getAllOrganization = async () => {
     try{
